# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * OkSchool API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Document {
  id: number;
  title: string;
  description: string;
  /** One of: notes, investigatory-projects, question-papers, free-book-pdfs, practical-files-class-12 */
  category: string;
  /** pdf or word */
  fileType: string;
  isFree: boolean;
  /** @nullable */
  price?: number | null;
  /** @nullable */
  fileUrl?: string | null;
  /** @nullable */
  wordFileUrl?: string | null;
  /** @nullable */
  thumbnailUrl?: string | null;
  viewCount: number;
  pdfDownloads: number;
  wordDownloads: number;
  createdAt: string;
}

export interface DocumentInput {
  /** @minLength 1 */
  title: string;
  description: string;
  category: string;
  fileType: string;
  isFree: boolean;
  /** @nullable */
  price?: number | null;
  /** @nullable */
  fileUrl?: string | null;
  /** @nullable */
  wordFileUrl?: string | null;
  /** @nullable */
  thumbnailUrl?: string | null;
}

export interface DocumentPatch {
  /** @minLength 1 */
  title?: string;
  description?: string;
  category?: string;
  fileType?: string;
  isFree?: boolean;
  /** @nullable */
  price?: number | null;
  /** @nullable */
  fileUrl?: string | null;
  /** @nullable */
  thumbnailUrl?: string | null;
}

export interface DownloadEventInput {
  /** pdf or word */
  fileType: string;
}

export interface AmazonProduct {
  id: number;
  title: string;
  description: string;
  affiliateUrl: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  price?: string | null;
  createdAt: string;
}

export interface AmazonProductInput {
  /** @minLength 1 */
  title: string;
  description: string;
  affiliateUrl: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  price?: string | null;
}

export interface AmazonProductPatch {
  title?: string;
  description?: string;
  affiliateUrl?: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  price?: string | null;
}

export interface MockTest {
  id: number;
  title: string;
  description: string;
  subject: string;
  /** Duration in minutes */
  duration: number;
  questionCount: number;
  createdAt: string;
}

export interface Question {
  id?: number;
  question: string;
  /** @nullable */
  questionImage?: string | null;
  optionA: string;
  optionB: string;
  optionC: string;
  optionD: string;
  /** A, B, C, or D */
  correctAnswer: string;
  /** @nullable */
  solution?: string | null;
}

export interface MockTestInput {
  /** @minLength 1 */
  title: string;
  description: string;
  subject: string;
  duration: number;
  questions?: Question[];
}

export interface MockTestDetail {
  id: number;
  title: string;
  description: string;
  subject: string;
  duration: number;
  questionCount: number;
  createdAt: string;
  questions: Question[];
}

export interface SectionCount {
  category: string;
  count: number;
}

export interface AdminStats {
  totalViews: number;
  pdfDownloads: number;
  wordDownloads: number;
  totalDocuments: number;
  recentDocuments: Document[];
  sectionCounts: SectionCount[];
}

export interface SearchResults {
  documents: Document[];
  amazonProducts: AmazonProduct[];
  mockTests: MockTest[];
}

export type ListDocumentsParams = {
category?: string;
q?: string;
};

export type GlobalSearchParams = {
q: string;
};