# lib/api-zod/src/generated/types/mockTestInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * OkSchool API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Question } from './question';

export interface MockTestInput {
  /** @minLength 1 */
  title: string;
  description: string;
  subject: string;
  duration: number;
  questions?: Question[];
}